Add environment dict, no checking yet.
The environment dict will be used to check whether a function or
variable has been declared before it is first used, and if not, then an
error message can be printed. It also works to link together the use and
the declaration of a function or variable for future use when creating
the assembly code.